producer/requirements.txt
pydicom
numpy
Pillow
confluent_kafka
pymongo